scrpy01/pyproject.toml
[tool.poetry]
name = "scrpy01"
version = "1.0.0"
description = "Hardcoded Credentials."
authors = ["Ande Sec <cyseceveryday@gmail.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.13"
Flask = "~3.1.0"
mysql-connector-python= "~9.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"